+++ AVACore/SocketKernel.js
@@ -1,5 +1,5 @@
 import socketIO from "socket.io";
-import { UUID } from "./Util";
+import { UUID } from "../AVAFoundation/AVAUtil";
 
 
 /**