+++ index.js
@@ -1,23 +1,22 @@
 // AVAFoundation
-export const AVAPushNotification = require("./AVAFoundation/AVAPushNotification.js").default;
-export const AVARecordZoneType = require("./AVAFoundation/AVARecordZoneType.js").default;
-export const AVALocalisation = require("./AVAFoundation/AVALocalisation.js").default;
-export const AVAQueryBuilder = require("./AVAFoundation/AVAQueryBuilder.js").default;
-export const AVAEnvironment = require("./AVAFoundation/AVAEnvironment.js").default;
-export const AVARecordZone = require("./AVAFoundation/AVARecordZone.js").default;
-export const AVAValidator = require("./AVAFoundation/AVAValidator.js").default;
-export const AVADatabase = require("./AVAFoundation/AVADatabase.js").default;
-export const AVAStorage = require("./AVAFoundation/AVAStorage.js").default;
-export const AVAMailer = require("./AVAFoundation/AVAMailer.js").default;
-export const AVAError = require("./AVAFoundation/AVAError.js").default;
-
-// Utilities
-export const Util = require("./AVACore/Util.js");
-export const CoreUtil = require("./AVACore/CoreUtil.js");
-export const _ = require("./AVAFoundation/AVALocalisation.js").translate;
+export const AVAPushNotification = require("./AVAFoundation/AVAPushNotification").default;
+export const AVARecordZoneType = require("./AVAFoundation/AVARecordZoneType").default;
+export const AVALocalisation = require("./AVAFoundation/AVALocalisation").default;
+export const AVAQueryBuilder = require("./AVAFoundation/AVAQueryBuilder").default;
+export const AVAEnvironment = require("./AVAFoundation/AVAEnvironment").default;
+export const AVARecordZone = require("./AVAFoundation/AVARecordZone").default;
+export const AVAValidator = require("./AVAFoundation/AVAValidator").default;
+export const AVADatabase = require("./AVAFoundation/AVADatabase").default;
+export const AVAStorage = require("./AVAFoundation/AVAStorage").default;
+export const AVAMailer = require("./AVAFoundation/AVAMailer").default;
+export const AVAError = require("./AVAFoundation/AVAError").default;
+export const AVAUtil = require("./AVAFoundation/AVAUtil");
 
 // Super components
-export const AVAViewController = require("./AVAFoundation/AVAViewController.js").default;
-export const AVAController = require("./AVAFoundation/AVAController.js").default;
-export const AVAMiddleware = require("./AVAFoundation/AVAMiddleware.js").default;
-export const AVAModel = require("./AVAFoundation/AVAModel.js").default;
+export const AVAViewController = require("./AVAFoundation/AVAViewController").default;
+export const AVAController = require("./AVAFoundation/AVAController").default;
+export const AVAMiddleware = require("./AVAFoundation/AVAMiddleware").default;
+export const AVAModel = require("./AVAFoundation/AVAModel").default;
+
+// Other
+export const _ = require("./AVAFoundation/AVALocalisation").translate;