+++ CHANGELOG.md
@@ -5,12 +5,14 @@ View detailed changes in our [Trello board](https://trello.com/b/EPJJJbYH/avalan
 
 ## [Unreleased]
 
-## [0.4.48] - 2019-11-17
+## [0.4.50] - 2019-11-17
 ### Added
 - Added AVAPushNotification class for iOS remote notifications.
 
 ### Changes
-- Started implementation of ESM. (A lot of refactoring)
+- Standardized AVAUtil & CoreUtil.
+- Updated package namespaces (AVAFoundation & AVACore).
+- Started implementation of ESM (A lot of refactoring).
 
 ## [0.4.31] - 2019-11-16
 ### Added